% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure.R
\name{age_structure}
\alias{age_structure}
\title{Run demo app}
\usage{
age_structure(ui = age_structure_ui, server = age_structure_server)
}
\arguments{
\item{ui}{Shiny UI defintion}

\item{server}{Shiny server function}
}
\description{
Run demo app
}
