% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_receivables.R
\name{create_receivables}
\alias{create_receivables}
\title{Create a data.frame of random receivables}
\usage{
create_receivables(seed = 80402)
}
\arguments{
\item{seed}{Random seed}
}
\description{
Create a data.frame of random receivables
}
